+++ tests/test_cli.py
@@ -129,6 +129,18 @@ def test_show_release_raw(changelog: str, capsys: pytest.CaptureFixture):
     )
 
 
+def test_show_release_does_not_exist(changelog: str, capsys: pytest.CaptureFixture):
+    # Requested release does not exist.
+    with pytest.raises(SystemExit) as cm:
+        cli(["show", "11.0.0", changelog])
+    assert cm.value.code == 3
+
+    captured = capsys.readouterr()
+
+    assert captured.err.strip() == f"{changelog} does not contain release 11.0.0."
+    assert captured.out == ""
+
+
 def test_create_release_automatic_version(
     changelog: str, capsys: pytest.CaptureFixture
 ):
@@ -174,3 +186,18 @@ def test_create_release_specific_version(changelog: str, capsys: pytest.CaptureF
 
     assert captured.err == ""
     assert captured.out.strip() == "3.2.1"
+
+
+def test_missing_command(capsys: pytest.CaptureFixture):
+    with pytest.raises(SystemExit) as cm:
+        cli([])
+    assert cm.value.code == 2
+
+    captured = capsys.readouterr()
+
+    assert (
+        captured.err.strip()
+        == """usage: keepachangelog [-h] [-v] {show,release} ...
+keepachangelog: error: the following arguments are required: {show,release}"""
+    )
+    assert captured.out == ""