+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
 - Do not release same version twice.
+- Show helpful error message if `show <release>` does not find requested `release`.
 
 ## [2.0.0] - 2024-06-14
 ### Removed